+++ pkg/scalers/authentication/authentication_types.go
@@ -12,6 +12,4 @@ const (
 	TLSAuthType Type = "tls"
 	// BearerAuthType is a auth type using a bearer token
 	BearerAuthType Type = "bearer"
-	// JwtAuthType is a auth type using a jwt token
-	JwtAuthType Type = "jwt"
 )