+++ pkg/scalers/metrics_api_scaler.go
@@ -52,9 +52,9 @@ type metricsAPIScalerMetadata struct {
 	key       string
 	ca        string
 
-	// jwt
-	enableJwtAuth bool
-	jwtToken      string
+	// bearer
+	enableBearerAuth bool
+	bearerToken      string
 }
 
 const (
@@ -163,13 +163,13 @@ func parseMetricsAPIMetadata(config *ScalerConfig) (*metricsAPIScalerMetadata, e
 
 		meta.key = config.AuthParams["key"]
 		meta.enableTLS = true
-	case authentication.JwtAuthType:
-		if len(config.AuthParams["jwtToken"]) == 0 {
-			return nil, errors.New("no jwtToken provided")
+	case authentication.BearerAuthType:
+		if len(config.AuthParams["token"]) == 0 {
+			return nil, errors.New("no token provided")
 		}
 
-		meta.jwtToken = config.AuthParams["jwtToken"]
-		meta.enableJwtAuth = true
+		meta.bearerToken = config.AuthParams["token"]
+		meta.enableBearerAuth = true
 	default:
 		return nil, fmt.Errorf("err incorrect value for authMode is given: %s", authMode)
 	}
@@ -317,12 +317,12 @@ func getMetricAPIServerRequest(meta *metricsAPIScalerMetadata) (*http.Request, e
 		}
 
 		req.SetBasicAuth(meta.username, meta.password)
-	case meta.enableJwtAuth:
+	case meta.enableBearerAuth:
 		req, err = http.NewRequest("GET", meta.url, nil)
 		if err != nil {
 			return nil, err
 		}
-		req.Header.Add("Authorization", meta.jwtToken)
+		req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", meta.bearerToken))
 	default:
 		req, err = http.NewRequest("GET", meta.url, nil)
 		if err != nil {